third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-img-element/list-of-available-images-does-not-coalesce-in-flight-requests.sub.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] list of available images does not coalesce in-flight requests
  assert_equals: The server should have seen exactly two requests, since the second image request above did not coalesce with the first in-flight one expected "2" but got "1"
Harness: the test ran to completion.
